[refactor] Create a package for the NameChecker in pylint.checker.base
